-modified some relation, add the possibility to add a recurrence in a calendar's event , move the logich on services and fix the controller; - Add a class Recurrence Manager and an enum RecurrenceType The project is now running

+++ src/main/java/DevelHope/Calendar/service/CalendarioService.java
@@ -1,29 +1,44 @@
 package DevelHope.Calendar.service;
 
 import DevelHope.Calendar.entity.Calendario;
+import DevelHope.Calendar.entity.Utente;
 import DevelHope.Calendar.repository.CalendarioRepository;
+import DevelHope.Calendar.repository.UtenteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class CalendarioService {
-
-    CalendarioRepository calendarioRepository;
-
     @Autowired
-    public CalendarioService(CalendarioRepository calendarioRepository) {
-        this.calendarioRepository = calendarioRepository;
-    }
+    CalendarioRepository calendarioRepository;
+    @Autowired
+    UtenteRepository utenteRepository;
 
-    public void createCalendario(Calendario calendario) {
-        calendarioRepository.save(calendario);
-    }
 
-    public void deleteCalendario(long id) {
-        calendarioRepository.deleteById(id);
+
+    public Calendario createCalendar(long userID, Calendario calendar) throws Exception{
+        if (utenteRepository.findById(userID).isPresent()){
+            Utente utente = utenteRepository.findById(userID).orElse(null);
+
+            calendar.setUtente(utente);
+
+        } else {
+            throw new Exception(String.format("User with ID %s not exist", userID));
+        }
+
+        return calendarioRepository.save(calendar);
+    }
+    public String deleteCalendario(long id) throws Exception {
+        if (calendarioRepository.findById(id).isPresent()) {
+            calendarioRepository.deleteById(id);
+            return String.format("Calendar id n° %S deleted", id);
+        } else {
+            throw new Exception(String.format("Calendar with ID %s not found", id));
+        }
     }
 
     public Calendario updateCalendario(long id, Calendario calendario) {
@@ -41,8 +56,8 @@ public class CalendarioService {
                     if (calendario.getDataCreazioneCalendario() != null) {
                         calendarioEsistente.setData(calendario.getDataCreazioneCalendario());
                     }
-                    if (calendario.getUtenti() != null) {
-                        calendarioEsistente.setUtenti(calendario.getUtenti());
+                    if (calendario.getUtente() != null) {
+                        calendarioEsistente.setUtente(calendario.getUtente());
                     }
                     return calendarioRepository.save(calendarioEsistente);
                 })
@@ -53,7 +68,13 @@ public class CalendarioService {
         return calendarioRepository.findById(id);
     }
 
-    public List<Calendario> vediCalendari() {
-        return calendarioRepository.findAll();
+    public List<Calendario> vediCalendariByUser(long userID) throws Exception {
+        Optional<Utente> utente = utenteRepository.findById(userID);
+
+        if (utente.isPresent()) {
+            return new ArrayList<>(utente.get().getCalendario());
+        } else {
+            throw new Exception(String.format("User with ID %s don't have calendars", userID));
+        }
     }
 }
