+++ src/main/java/DevelHope/Calendar/service/EventoService.java
@@ -1,29 +1,52 @@
 package DevelHope.Calendar.service;
 
+import DevelHope.Calendar.entity.Calendario;
 import DevelHope.Calendar.entity.Evento;
+import DevelHope.Calendar.entity.Utente;
+import DevelHope.Calendar.repository.CalendarioRepository;
 import DevelHope.Calendar.repository.EventoRepository;
+import DevelHope.Calendar.repository.UtenteRepository;
+import DevelHope.Calendar.util.RecurrenceManager;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class EventoService {
-
-    EventoRepository eventoRepository;
-
     @Autowired
-    public EventoService(EventoRepository eventoRepository) {
-        this.eventoRepository = eventoRepository;
+    EventoRepository eventoRepository;
+    @Autowired
+    CalendarioRepository calendarioRepository;
+    @Autowired
+    UtenteRepository utenteRepository;
+
+    public Evento createEvent(long calendarioID, Evento event, LocalDateTime startTime, int duration) throws Exception {
+
+        RecurrenceManager recurrenceManager = new RecurrenceManager();
+
+        if (calendarioRepository.findById(calendarioID).isPresent()) {
+            Calendario calendario = calendarioRepository.findById(calendarioID).orElse(null);
+
+            event.setCalendario(calendario);
+
+            recurrenceManager.setRecurrenceTimeAndDuration(event, startTime, duration);
+        } else {
+            throw new Exception(String.format("Calendar with ID %s not found", calendarioID));
+        }
+
+        return eventoRepository.save(event);
     }
 
-    public void createEvento(Evento evento) {
-        eventoRepository.save(evento);
-    }
-
-    public void deleteEvento(long id) {
-        eventoRepository.deleteById(id);
+    public String deleteEvento(long id) throws Exception {
+        if (eventoRepository.findById(id).isPresent()) {
+            eventoRepository.deleteById(id);
+            return String.format("Event id n° %S deleted", id);
+        } else {
+            throw new Exception(String.format("Event with ID %s not found", id));
+        }
     }
 
     public Evento updateEvento(long id, Evento evento) {
@@ -34,9 +57,6 @@ public class EventoService {
                     }
                     if (evento.getDescrizione() != null) {
                         eventoEsistente.setDescrizione(evento.getDescrizione());
-                    }
-                    if (evento.getColor() != null) {
-                        eventoEsistente.setColor(evento.getColor());
 
                     }
                     if (evento.getStartTime() != null) {
@@ -45,23 +65,41 @@ public class EventoService {
                     if (evento.getEndTime() != null) {
                         eventoEsistente.setEndTime(evento.getEndTime());
                     }
-                    if (evento.getInvitati() != null) {
-                        eventoEsistente.setInvitati(evento.getInvitati());
-                    }
-                    if (evento.getCalendari() != null) {
-                        eventoEsistente.setCalendari(evento.getCalendari());
-                    }
 
                     return eventoRepository.save(eventoEsistente);
                 })
                 .orElse(null);
     }
 
-    public Optional<Evento> vediEvento(long id) {
-        return eventoRepository.findById(id);
-    }
+    public List<Evento> vediEvento(long calendarioID) throws Exception {
+        if (calendarioRepository.findById(calendarioID).isPresent()) {
+            Calendario calendar = calendarioRepository.findById(calendarioID).orElse(null);
 
-    public List<Evento> vediEvento() {
-        return eventoRepository.findAll();
+            List<Evento> eventi;
+            eventi = calendar.getEventi();
+
+            return eventi;
+
+        } else {
+            throw new Exception(String.format("Calendar with ID %s not exist", calendarioID));
+        }
+    }
+    @Transactional
+    public String inviteUser(long eventoID, long utenteID) throws Exception{
+
+        Evento event = eventoRepository.findById(eventoID).get();
+        Utente user = utenteRepository.findById(utenteID).get();
+
+        if (event == null || user == null){
+
+            throw new Exception("Event or user not found");
+        } else {
+            event.getInvitati().add(user);
+            user.getEventi().add(event);
+            eventoRepository.save(event);
+            utenteRepository.save(user);
+            return "User successfully invited ";
+        }
+
     }
 }
